Hoist log_warnings test fixture into shared conftest
Problem: A near-identical log_warnings pytest fixture was duplicated in
two test files (libs/mngr/imbue/mngr/cli/common_opts_test.py and
libs/mngr/imbue/mngr/config/loader_test.py). The cli copy had a
defensive try/except around logger.remove (to tolerate setup_logging
clearing handlers), while the config copy did not -- so the two
implementations had quietly drifted.

Fix: Move log_warnings into libs/mngr/imbue/mngr/conftest.py (the
nearest common-ancestor conftest) using the more-defensive variant, and
drop the duplicate fixtures (and now-unused Generator/logger imports)
from both test files. Pytest fixture lookup makes the conftest version
visible to all tests under imbue/mngr/.
